Add analytical model
This model compute the analytical expression for the Helmholtz
transmission problem in acoustics for a plane wave being scattered
by a sphere (centered at the origin).

It is aimed to validate the OptimUS numerical simulations.

The class checks for the validity of the parameters. Specifically
it checks for:
- the geometry being a sphere (though, not any kind of sphere, ie.
  does not check for Regular sphere)
- the source being a plane wave
- only one bounded domain

The model solves the problem by computing the coefficients of the
series for the scattered exterior and interior fields. It does not
check or alert division by zero and does not have a tolerance build in
(in my experience, setting a tolerance provided poor results).