Add analytical model to visualization pipeline
For specific functionality the analytical model is
derived to its own visualization scheme (compatible
with the OptimUS). Basically it recycles everything
with the exception of the compute_pressure_fields function
that has its own analytical version. The modified methods and
functions in the OptimUS library only derived the analytical
model to its specific functions.